fix(clips): keep reference controller bounded (#1951)
